--- a/cpp/benchmarks/core/CMakeLists.txt
+++ b/cpp/benchmarks/core/CMakeLists.txt
@@ -1,5 +1,6 @@
 target_sources(benchmarks PRIVATE
     Hashmap.cpp
+    NNS.cpp
     Reduction.cpp
     Zeros.cpp
 )