NNS performance 
0.02 voxel size on test_data/fragment pointcloud. 
search radius: 0.04, 0.07
max_nn: 1, 30

### [Updated]
### Master
i7-7700HQ, GTX 1050Ti 4GB, 8GB RAM
```
--------------------------------------------------------------------------
Benchmark                                Time             CPU   Iterations
--------------------------------------------------------------------------
HybridSearch/[0.07 | 30] CPU          14.3 ms         14.2 ms           47
HybridSearch/[0.07 | 30] CUDA         21.2 ms         21.2 ms           33
HybridSearch/[0.04 | 30] CPU          1.50 ms         1.50 ms          451
HybridSearch/[0.04 | 30] CUDA        0.807 ms        0.807 ms          860
HybridSearch/[0.07 | 1] CPU           13.9 ms         13.9 ms           48
HybridSearch/[0.07 | 1] CUDA          1.77 ms         1.77 ms          392
KnnSearch/[30] CPU                    9.89 ms         5.69 ms          102
KnnSearch/[30] CUDA                   95.2 ms         95.2 ms            7
KnnSearch/[1] CPU                     9.88 ms         4.59 ms          149
KnnSearch/[1] CUDA                    84.6 ms         84.6 ms            8
LegacySearch/[30] KNN                 26.4 ms         26.4 ms           27
LegacySearch/[1] KNN                  13.3 ms         13.3 ms           53
LegacySearch/[0.07 | 30] Hybrid       26.4 ms         26.4 ms           27
LegacySearch/[0.07 | 1] Hybrid        13.3 ms         13.3 ms           53
```
### PR 3237
i7-7700HQ, GTX 1050Ti 4GB, 8GB RAM
```
--------------------------------------------------------------------------
Benchmark                                Time             CPU   Iterations
--------------------------------------------------------------------------
HybridSearch/[0.07 | 30] CPU          14.4 ms         14.4 ms           48
HybridSearch/[0.07 | 30] CUDA         4.86 ms         4.86 ms          128
HybridSearch/[0.04 | 30] CPU          1.55 ms         1.55 ms          446
HybridSearch/[0.04 | 30] CUDA        0.750 ms        0.750 ms          928
HybridSearch/[0.07 | 1] CPU           14.1 ms         14.1 ms           49
HybridSearch/[0.07 | 1] CUDA          2.42 ms         2.42 ms          294
KnnSearch/[30] CPU                    10.0 ms         5.52 ms          100
KnnSearch/[30] CUDA                   79.5 ms         79.5 ms            9
KnnSearch/[1] CPU                     10.3 ms         5.47 ms          147
KnnSearch/[1] CUDA                    60.3 ms         60.3 ms           12
LegacySearch/[30] KNN                 26.0 ms         26.0 ms           27
LegacySearch/[1] KNN                  13.4 ms         13.4 ms           52
LegacySearch/[0.07 | 30] Hybrid       26.5 ms         26.5 ms           27
LegacySearch/[0.07 | 1] Hybrid        13.5 ms         13.5 ms           52
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/intel-isl/open3d/3688)
<!-- Reviewable:end -->