@reyanshsolis 2021-07-06
@junha-l may you please once verify if the benchmark conditions are acceptable.

We may want to take a look at the possible reasons for the unexpected performance.