@junha-l 2021-07-06
It seems problematic.
HybridSearch and KnnSearch don't downsample point cloud whereas the legacy does.
I pushed the patch for this (https://github.com/intel-isl/Open3D/pull/3688/commits/06087039ec1f53d168706eb266b2cbd502d2bdb0).

Here is the new result on my machine
```
------------------------------------------------------------------------
Benchmark                              Time             CPU   Iterations
------------------------------------------------------------------------
HybridSearch/[0.07 | 30] CPU        8.27 ms         7.67 ms           86
HybridSearch/[0.07 | 30] CUDA       1.86 ms         1.85 ms          330
HybridSearch/[0.04 | 30] CPU        1.07 ms         1.06 ms          729
HybridSearch/[0.04 | 30] CUDA      0.347 ms        0.341 ms         2315
HybridSearch/[0.07 | 1] CPU         7.84 ms         7.73 ms           94
HybridSearch/[0.07 | 1] CUDA       0.585 ms        0.585 ms         1161
KnnSearch/[30] CPU        5.49 ms         2.86 ms          365
KnnSearch/[30] CUDA       17.3 ms         17.3 ms           39
KnnSearch/[1] CPU         5.53 ms         2.94 ms          294
KnnSearch/[1] CUDA        13.6 ms         13.6 ms           48
LegacySearch/[30] KNN                 17.9 ms         17.9 ms           41
LegacySearch/[1] KNN                  8.80 ms         8.78 ms           80
LegacySearch/[0.07 | 30] Hybrid       17.6 ms         17.6 ms           38
LegacySearch/[0.07 | 1] Hybrid        9.19 ms         9.16 ms           82
```