@junha-l 2021-07-06
@reyanshsolis 
Could you check that there is the same problem in https://github.com/intel-isl/Open3D/pull/3650?