@reyanshsolis 2021-07-06
@junha-l Thanks, issues are resolved, even for `EstimateX`. 
It was a blunder from my side.
 
KNN CUDA might still need some testing, but I am going with HybridSearch for now, for EstimateX ops, to unblock the pr.