@junha-l 2021-07-06
@reyanshsolis 
Great!
As you said, now it seems that KNN CUDA performance is problematic. 
I'll take a look into it. Thanks